환자 정보 관리 모달 기능 구현
- PatientInformation 컴포넌트 생성 (환자 등록/수정/조회 모달)
- PatientRegistration을 PatientInformation 사용하도록 리팩토링
- 환자 삭제 기능 추가
- 환자 검색 기능 추가
- 테이블 헤더 '수정과 조회'로 변경
- 버튼 순서 변경 (수정  조회)
- 조회 아이콘을 FileChartColumnIncreasing으로 변경
- showHeader prop 추가로 헤더 표시 제어 가능
